Updated Hosting docs to include advice for Cloudways Permissions.
Cloudways has unusual permissions issues due to them not allowing applications or operations to run as Sudo. This occasionally results in weird situations where file permissions are incorrect and cannot be easily fixed without logging into their Admin interface. Resetting the permissions on their systems is a very commonly needed activity due to the security constraints of their stack.

Added a a bit of catch all documentation to reflect this.